… ChangeLog.md
… cpp/order/Steinhardt.h
… doc/source/reference/credits.rst
… freud/_order.pxd
… freud/order.pyx
… freud/plot.py
… tests/test_order_Steinhardt.py