<!-- Provide a general summary of your changes in the Title above. -->

## Description
<!-- Describe your changes in detail. -->
This adds support for multiple `l` values in the `Steinhardt` class. I also introduce significant speed ups through three changes
- Only construct `fsph::PointSPHEvaluator` once per thread per `baseCompute` (the construction is somewhat costly)
- Reduce allocations of `std::vector` this surprisingly comprised a decent bit of the time computing.
- Reduce calls to `ManagedArray::getIndex` or the use of the `operator()`. This reduces `std::vector` allocations further and reduces computation by accessing the data directly through `operator[]`.

The changes are implemented in a backwards compatible way. Currently if the user only provides a single `l` value, everything behaves as before, but if they provide multiple, we return a list of parameters for all the order parameters provided by the compute.

Edit: The new approach uses a multidimensional `ManagedArray` and NumPy arrays.

## Motivation and Context
<!-- Why is this change required? What problem does it solve? -->
<!-- Replace ??? with the issue number that this pull request resolves, if applicable. -->
Resolves: #747

This is desired since using multiple `Steinhardt` objects is significantly slower than using `LocalDescriptors` and a custom Numba solution to compute the Steinhardt order parameters (even in serial).

Here is a brief benchmark comparing the new code to the old and the custom code mentioned in issue #747. For more details see the comment below.

| Name | Time| Commit |
| ------- | ------- | -------- |
| New Code (l = [0, 12]) | ~169ms | 345a0dd |
| New Code with `ManagedArray`s | ~330ms | 8371e9a |
| Custom Code (l = [0, 12]) | ~334ms | NA |
| New Code (l=6) | ~57.7ms | 345a0dd |
| Old Code (l=6) | ~190ms | f9599480 |

\* The commit hash for "Old Code" may be wrong. I just used the HEAD of the master branch for comparison.

Edit: Due to rebasing the new code commit hash does not exist on the branch any more.

## How Has This Been Tested?
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran to
     see how your changes affect other areas of the code, etc. -->
I added a test like the one that checks the validity of the spherical harmonic computation using multiple `l` values on a single compute.

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds or improves functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation improvement (updates to user guides, docstrings, or developer docs)

## Checklist:
<!-- Put an `x` in all the boxes that apply. If you're unsure about any of
     these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [**CONTRIBUTING**](https://github.com/glotzerlab/freud/blob/master/CONTRIBUTING.md) document.
- [x] My code follows the code style of this project.
- [x] I have updated the documentation (if relevant).
- [x] I have added tests that cover my changes (if relevant).
- [x] All new and existing tests passed.
- [x] I have updated the [credits](https://github.com/glotzerlab/freud/blob/master/doc/source/reference/credits.rst).
- [x] I have updated the [Changelog](https://github.com/glotzerlab/freud/blob/master/ChangeLog.md).