@b-butler 2021-05-14
Some rough bench-marking on a FCC system of 4000 particles with some noise. The benchmarks were run on my laptop which has 4 cores with hyperthreading (an i5-8250U for those interested). I did not bother to close all my programs to get the cleanest reading. I ran the benchmarks in a Jupyter notebook with the `%%timeit` magic command a few times and made sure they were sufficiently close. All neighborlist were precomputed to prevent this from having an effect. This is for the standard computation (not averaged or `Wl`). 

I am going to put this at the top of the PR but wanted to record when I added it as well.