@b-butler 2021-05-15
I changed to using `ManagedArray` objects over vectors of `ManagedArray` objects. I did notice it was slower about the same speed for the custom code benchmarked above and mentioned in #747.