@bdice 2021-05-15
> I changed to using `ManagedArray` objects over vectors of `ManagedArray` objects. I did notice it was slower about the same speed for the custom code benchmarked above and mentioned in #747.

We don’t have to make this change if there is a measurable performance difference. I am a little surprised, however. Can you share your benchmarks with commit ids for each version tested? I can review the implementation and see what might be wrong with the performance if you agree that this design choice (one large array) is preferable.