@b-butler 2021-05-17
@bdice I added commit hashes. I may not get back to this today to fix the tests, but it works locally for me, and feel free to test our the commits locally to see if you find what might be causing the slow down. I think the singular NumPy array is nice if we can do it without sacrificing significant performance.