@vyasr 2021-05-17
I opted for simplicity, generality, and safety in the implementation of `ManagedArray` over performance wherever there was an obvious choice to be made. In code where we have higher dimensional arrays (like PMFTs) or in general in any code where we index into ManagedArray objects in nested loops, those choices have noticeable performance penalties. I've worked around these before, but I can work with @b-butler to resolve this more fundamentally if we think it's worthwhile. There will by necessity be some tradeoffs to be made if we choose that, at the very least in the form of more code to maintain if we try to maintain both slow/safe and fast/unsafe code paths, but if the bottlenecks are hindering analysis then maybe it's worth it.