@b-butler 2021-05-18
@vyasr  I think this was the right choice especially since generally we don't need this high performance multidimensional access. I am not sure if this case would warrant creating unsafe paths for performance reasons, but if you have thoughts on improvements on my usage please suggest them.