@bdice 2021-05-28
I see the same performance between commit 345a0dd7, which used a vector of ManagedArrays, and commit 13e856a, which uses a 2D Managed Array. Both commits are about 2x faster than the `master` branch for large systems and one value of `l`.

345a0dd7:
![image](https://user-images.githubusercontent.com/3943761/120041250-8970bd00-bfcd-11eb-9f55-b483429cafab.png)

13e856a:
![image](https://user-images.githubusercontent.com/3943761/120041283-955c7f00-bfcd-11eb-8029-d287e4faddef.png)