@b-butler 2021-06-01
@vyasr most of your comments on loops I disagree with shifting the loop bounds would simplify the code, since the different arrays are indexed differently. In all cases at least two four loop variables would need to be defined and incremented. If you prefer multiple indices to be created in a for loop, then I will make the change.