@vyasr 2021-06-04
> @vyasr most of your comments on loops I disagree with shifting the loop bounds would simplify the code, since the different arrays are indexed differently. In all cases at least two four loop variables would need to be defined and incremented. If you prefer multiple indices to be created in a for loop, then I will make the change.

Caught up with @b-butler briefly offline. I agree that in many cases you still need to maintain multiple indices, so just shifting the bounds may not be so helpful. I am a little hesitant to rely on the compiler to optimize out the multiple computations of the loop index, so I have a slight preference for computing it once and storing it to make the compiler's job easier, but aside from that I don't think any changes are necessary on the indexing front. I'll leave it to @b-butler to make the call there, then please re-request my review and we can wrap this up.